pyproject.toml
[project]
name = "findora"
version = "0.1.0"
description = "Personal-Use AI Outfit Recommendation Agent"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "playwright>=1.44.0",
    "chromadb>=0.5.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "apscheduler>=3.10.0",
    "peewee>=3.17.0",
    "python-dotenv>=1.0.0"
]

[tool.setuptools]
packages = ["app"]
